Split out ParserCommon to Type and Builtin.
