+++ plutus-core/plutus-core.cabal
@@ -190,6 +190,8 @@ library
         PlutusCore.Evaluation.Machine.ExBudgetingDefaults
         PlutusCore.InlineUtils
         PlutusCore.Parser.ParserCommon
+        PlutusCore.Parser.Type
+        PlutusCore.Parser.Builtin
         PlutusCore.Pretty.Classic
         PlutusCore.Pretty.ConfigName
         PlutusCore.Pretty.ConfigName
